requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
requests
fastapi
uvicorn
websockets
psycopg2-binary
pymongo